Drop the round-about Win32 console attachment switcheroo
Background information:

On Win32, a binary is compiled targeting a specific "kernel subsystem".
Two prominent subsystems are called "console" (really "CUI") and
"windows" ("GUI"), respectively (cf [1]).

The "CUI" subsystem is used for so-called "character-mode" applications.
The primary feature discerning console-mode from windows-mode, for the
purposes of this discussion, is that the operating system takes care of
providing a console for console-mode applications. If a console is
already present, the freshly launched application will be connected to
this parent console (redirection stdin, stdout, stderr). If no console
is present, the operating system will launch a console host before
launching the application itself.

This console management is not done for windows-mode applications.
A windows-mode application can of course also peruse a console. But it
has to itself take care of allocating and attaching to the console.

When an interpreter is to run both graphical and console-oriented
scripts on Windows, it thus faces a inevitable problem: CUI or GUI mode?

Using a console-mode binary for running GUI scripts results in an ugly
"console flicker". As console-mode asks the Windows OS to take care of
allocating the console, it will _always_ do so, before even passing over
startup to the interpreter.

Using a GUI-mode binary for running console-oriented scripts results in
woes with process coordination and stdio redirection, especially when
combined with interactive input and a pre-existing console (i.e. the
interpreter launched from inside an existing console host). GUI-mode
binaries immediately detach from a pre-existing console host when they
are launched. Even though they can later re-attach to a pre-existing
parent console, there is a time window when this is not yet set up. And
the missing parent/child relationship between the processes cause other
issues, such as the parent being unaware when the child has exited.

As Saphirion's (and later, Atronix) interpreter features both, a GUI
dialect and "encapping" (which bundles a Rebol interpreter along with a
script into a standalone executable), it also had to face the
console-mode vs GUI-mode dilemma. Their solution boils down to allowing
both modes as compilation targets, with roughly the following behaviour:

- When compiled as console-mode binary, the console is (re-)used
  properly. GUI scripts run with console-mode interpreters exhibit
  "console flicker".

- When compiled as GUI-mode binary:

    - Console visibility is explicitly manageable via SHOW-CONSOLE /
      HIDE-CONSOLE functions (provided by the "core" extension).

    - A console is allocated and/or attached to once the first output to
      a console (using eg PRINT) is attempted. This is toggleable via
      the CONSOLE-OUTPUT function (provided by the "core" extension).

    - If the interpreter is started from a parent console and no
      command-line arguments are present (not sure about the reason for
      the latter restriction), it re-forks itself to avoid any
      possibility of attaching to the parent console.

    - If running a script (encapped or directly), any console output is
      suppressed (using the `console-ouput false` equivalent) and only
      re-enabled right before the actual script is executed. (This was
      done in a %saphir-patches.r boot file, which got removed during
      the integrating Atronix branch into Ren/C.)

Overall, this is a well-intentioned. One of the main effects, in
practice, seem to be that GUI-mode binaries are useable for shipping
encapped GUI apps where "console flicker" is inappropriate, whereas the
console-mode binary is more widely applicable, if that "flicker" is not
an issue.

Finally, the actual implementation of this dual compilation-mode
suppport evolved over a series of attempts of getting basic stuff to
work nicely. The current final state was rather round-about and in
desparate need of consolidation and clean-up.

[1] http://blogs.msdn.com/b/oldnewthing/archive/2009/01/01/9259142.aspx
[2] http://www.tillett.info/2013/05/13/how-to-create-a-windows-program-that-works-as-both-as-a-gui-and-console-application/

Intended change:

As we currently don't support a GUI dialect and our only intended
compilation target on Windows is console-mode binaries, the remaining
code implementing the console attachment is removed.